+++ .eslintrc.js
@@ -4,6 +4,7 @@ module.exports = {
   root: true,
   env: {
     browser: true,
+    node: true
   },
   extends: [
     'eslint:recommended',