+++ config/prod.env.js
@@ -1,6 +1,0 @@
-'use strict'
-
-module.exports = {
-  NODE_ENV: '"production"',
-  HONEYBADGER_API_KEY: `"${process.env.HONEYBADGER_API_KEY}"`
-}