+++ test/unit/TestCanvasForProps.vue
@@ -1,6 +1,7 @@
 <template>
   <div>
     <test-component-with-props
+      ref="test-comp-with-props"
       :title="name"
       :count="total"
     />
@@ -12,16 +13,14 @@ import TestComponentWithProps from './TestComponentWithProps'
 export default {
   name: 'TestCanvasForProps',
   components: { TestComponentWithProps },
-  filters: {
-  },
   data () {
     return {
       name: 'Component 1',
-      total: -1
+      total: 10
     }
   },
   methods: {
-  },
+  }
 }
 </script>
 <style scoped>