+++ test/unit/TestComponentWithProps.vue
@@ -1,30 +1,29 @@
 <template>
-  <div
-    class="miniwolf"
-    @customEvent="blowup()"
-  >
+  <div class="miniwolf">
     <h1>{{ title }}</h1>
-    <div>{{ count | assertInRange }}</div>
+    <div>{{ assertInRange }}</div>
   </div>
 </template>
 
 <script>
 export default {
   name: 'TestComponentWithProps',
-  filters: {
-    assertInRange: function (value) {
-      if (value < 0) { throw Error('out of range') }
-      return value
-    }
-  },
   props: {
     title: {
       type: String,
-      required: true
+      default: ''
     },
     count: {
       type: Number,
-      required: true
+      default: -1
+    }
+  },
+  computed: {
+    assertInRange: function () {
+      if (this.count < 0) {
+        throw Error('out of range')
+      }
+      return this.count
     }
   }
 }